Implement string interning for WorldStmt
Cache references to strings used in names for statements. Modify
interner to return Rc<str> to avoid lifetime issues. WorldStmt is generic
over the AsRef<str> type, which is a string slice into the input when
returned by the parsers then immediately turned into a cached Rc<str>
by the interner to deduplicate and allow the backing input to be dropped.